RUM-2150 Add enriched resource
