+++ DatadogSessionReplay/Sources/Feature/RequestBuilders/ResourceRequestBuilder.swift
@@ -19,12 +19,12 @@ internal struct ResourceRequestBuilder: FeatureRequestBuilder {
     func request(for events: [Event], with context: DatadogContext) throws -> URLRequest {
         let decoder = JSONDecoder()
         let resources = try events.map { event in
-            try decoder.decode(CodableResource.self, from: event.data)
+            try decoder.decode(EnrichedResource.self, from: event.data)
         }
         return try createRequest(resources: resources, context: context)
     }
 
-    private func createRequest(resources: [CodableResource], context: DatadogContext) throws -> URLRequest {
+    private func createRequest(resources: [EnrichedResource], context: DatadogContext) throws -> URLRequest {
         var multipart = multipartBuilder
 
         let builder = URLRequestBuilder(
@@ -63,34 +63,6 @@ internal struct ResourceRequestBuilder: FeatureRequestBuilder {
 
     private func url(with context: DatadogContext) -> URL {
         customUploadURL ?? context.site.endpoint.appendingPathComponent("api/v2/replay")
-    }
-}
-
-internal struct CodableResource: Codable, Hashable, Resource {
-    internal struct Context: Codable, Equatable {
-        internal struct Application: Codable, Equatable {
-            let id: String
-        }
-        let type: String
-        let application: Application
-
-        init(_ applicationId: String) {
-            self.type = "resource"
-            self.application = .init(id: applicationId)
-        }
-    }
-    internal var identifier: String
-    internal var data: Data
-    internal var context: Context
-
-    internal init(resource: Resource, context: Context) {
-        self.identifier = resource.identifier
-        self.data = resource.data
-        self.context = context
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(identifier)
     }
 }
 #endif
