+++ DatadogSessionReplay/Sources/Writer/Models/EnrichedResource.swift
@@ -1,0 +1,37 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2019-Present Datadog, Inc.
+ */
+
+#if os(iOS)
+import Foundation
+
+internal struct EnrichedResource: Codable, Hashable, Resource {
+    internal struct Context: Codable, Equatable {
+        internal struct Application: Codable, Equatable {
+            let id: String
+        }
+        let type: String
+        let application: Application
+
+        init(_ applicationId: String) {
+            self.type = "resource"
+            self.application = .init(id: applicationId)
+        }
+    }
+    internal var identifier: String
+    internal var data: Data
+    internal var context: Context
+
+    internal init(resource: Resource, context: Context) {
+        self.identifier = resource.identifier
+        self.data = resource.data
+        self.context = context
+    }
+
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(identifier)
+    }
+}
+#endif
